# 04-API_SPEC.yaml
openapi: 3.1.0
...
info:
  title: ProofPay Application API
  version: 0.1.0
  description: 'ProofPay compiles the review queue between ''work submitted'' and ''payment released''
    into executable acceptance checks — and pays only when evidence passes.


    Phase 2 contract, not an implemented server. Read docs/00-PROJECT_CONTEXT.md and docs/02-PRD.md. One
    agency, one fixture, two sandbox recipients, USD, exactly three approved checks and three families.
    Examples/identifiers are illustrative unless execution evidence says otherwise. Cookie domain mutations
    require CSRF and Idempotency-Key; internal tool/runner credentials never enter model context. Paid
    requires a matched actual PayPal item SUCCESS.'
  license:
    name: MIT
    identifier: MIT
servers:
- url: http://localhost:8000
  description: Local Compose API; production cookies require HTTPS.
...
security:
- SessionCookie: []
paths:
...
  /api/briefs:
    get:
      operationId: list_briefs
      summary: list briefs
      description: ''
      tags:
      - Briefs
      x-roles:
      - owner
      x-requirement-ids:
      - FR-17
      responses:
        '200':
          description: Success
          headers:
            X-Request-ID:
              schema: *id001
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BriefList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
        default:
          $ref: '#/components/responses/Unexpected'
      security:
      - SessionCookie: []
      parameters:
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
    post:
      operationId: create_brief
      summary: create brief
      description: Create first immutable revision, task/demo-run binding, empty mandate container and
        zero-balance obligation. Validate supported scope and proposed terms; no release authority yet.
      tags:
      - Briefs
      x-roles:
      - owner
      x-requirement-ids:
      - FR-01
      - FR-17
      responses:
        '201':
          description: Success
          headers:
            X-Request-ID:
              schema: *id001
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Brief'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/Unexpected'
      security:
      - SessionCookie: []
      parameters:
      - $ref: '#/components/parameters/Csrf'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BriefCreate'
  /api/briefs/{brief_id}:
    get:
      operationId: get_brief
      summary: get brief
      description: ''
      tags:
      - Briefs
      x-roles:
      - owner
      x-requirement-ids:
      - FR-17
      responses:
        '200':
          description: Success
          headers:
            X-Request-ID:
              schema: *id001
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Brief'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
        default:
          $ref: '#/components/responses/Unexpected'
      security:
      - SessionCookie: []
      parameters:
      - name: brief_id
        in: path
        required: true
        schema: *id001
...
  /api/briefs/{brief_id}/compile:
    post:
      operationId: compile_brief
      summary: compile brief
      description: Queue compilation pinned to revision/digest. Ambiguous result blocks approval; no test
        code generated.
      tags:
      - Briefs
      x-roles:
      - owner
      x-requirement-ids:
      - FR-02
      - FR-18
      - NFR-17
      responses:
        '202':
          description: Durable work acknowledged; inspect referenced job/resource.
          headers:
            X-Request-ID:
              schema: *id001
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedCommand'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/Unexpected'
      security:
      - SessionCookie: []
      parameters:
      - name: brief_id
        in: path
        required: true
        schema: *id001
      - $ref: '#/components/parameters/Csrf'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompileRequest'
...
  /api/mandates:
    post:
      operationId: create_mandate
      summary: create mandate
      description: Populate the milestone existing mandate container with its first draft from a ready
        matching compilation. One mandate/obligation per task; additional drafts use the versions endpoint.
      tags:
      - Mandates
      x-roles:
      - owner
      x-requirement-ids:
      - FR-03
      - FR-19
      responses:
        '201':
          description: Success
          headers:
            X-Request-ID:
              schema: *id001
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MandateVersion'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/Unexpected'
      security:
      - SessionCookie: []
      parameters:
      - $ref: '#/components/parameters/Csrf'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MandateDraftCreate'
  /api/mandates/{mandate_id}:
    get:
      operationId: get_mandate
      summary: get mandate
      description: Contractors see only approved versions assigned to their recipient; owner sees permitted
        draft/history.
      tags:
      - Mandates
      x-roles:
      - owner
      - contractor
      x-requirement-ids:
      - FR-03
      - FR-20
      responses:
        '200':
          description: Success
          headers:
            X-Request-ID:
              schema: *id001
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mandate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
        default:
          $ref: '#/components/responses/Unexpected'
      security:
      - SessionCookie: []
      parameters:
      - name: mandate_id
        in: path
        required: true
        schema: *id001
...
  /api/mandates/{mandate_id}/versions/{version_id}/approve:
    post:
      operationId: approve_mandate
      summary: approve mandate
      description: Under shared authority lock, compare displayed digest/current pointer; freeze snapshot/checks/binding.
        Subsequent payload mutation forbidden.
      tags:
      - Mandates
      x-roles:
      - owner
      x-requirement-ids:
      - FR-03
      - FR-04
      - FR-19
      - FR-20
      responses:
        '200':
          description: Success
          headers:
            X-Request-ID:
              schema: *id001
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MandateVersion'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/Unexpected'
      security:
      - SessionCookie: []
      parameters:
      - name: mandate_id
        in: path
        required: true
        schema: *id001
      - name: version_id
        in: path
        required: true
        schema: *id001
      - $ref: '#/components/parameters/Csrf'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveRequest'
...
  /api/evidence/{bundle_id}:
    get:
      operationId: get_evidence_bundle
      summary: get evidence bundle
      description: ''
      tags:
      - Evidence
      x-roles:
      - owner
      - contractor
      x-requirement-ids:
      - FR-06
      - FR-07
      - FR-23
      responses:
        '200':
          description: Success
          headers:
            X-Request-ID:
              schema: *id001
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceBundle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
        default:
          $ref: '#/components/responses/Unexpected'
      security:
      - SessionCookie: []
      parameters:
      - name: bundle_id
        in: path
        required: true
        schema: *id001
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: proofpay_session
      description: Opaque session; role/agency resolved server-side.
    WorkerServiceToken:
      type: http
      scheme: bearer
      description: Application worker audience; job/stage/lease context checked. Never a PayPal token.
    RunnerServiceToken:
      type: http
      scheme: bearer
      description: Distinct runner audience; no app financial or AI access.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: *id001
      description: 'Required for domain mutations. Scope: agency, original principal, effective persona,
        canonical command family and key. Contractor delivery aliases share submit_delivery. Recheck authorization
        before replaying original result; changed digest gives 409. Retained across reset.'
    Csrf:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        maxLength: 128
      description: Session-bound token; required with cookie authentication on mutation.
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
        maxLength: 256
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
  responses:
    BadRequest:
      description: Malformed JSON/header or duplicate keys.
      headers:
        X-Request-ID:
          schema: *id001
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing/expired session or service credential.
      headers:
        X-Request-ID:
          schema: *id001
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Role, CSRF, service audience or stage denied.
      headers:
        X-Request-ID:
          schema: *id001
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource absent or inaccessible in agency/contractor scope.
      headers:
        X-Request-ID:
          schema: *id001
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Stale digest/version, idempotency conflict or workspace busy.
      headers:
        X-Request-ID:
          schema: *id001
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Invalid:
      description: Schema, scope, parameter, money or authority input invalid.
      headers:
        X-Request-ID:
          schema: *id001
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: Request limit reached; retry after the supplied delay.
      headers:
        X-Request-ID:
          schema: *id001
        Retry-After:
          schema:
            type: string
            maxLength: 80
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unavailable:
      description: Durable storage or required service unavailable.
      headers:
        X-Request-ID:
          schema: *id001
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unexpected:
      description: Unexpected error; use request_id without exposed secrets.
      headers:
        X-Request-ID:
          schema: *id001
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    PaymentTerms:
      type: object
      additionalProperties: false
      properties:
        recipient_ref:
          type: string
          enum:
          - contractor_maya
          - contractor_leo
        amount:
          $ref: '#/components/schemas/PositiveMoney'
        principal_cap:
          $ref: '#/components/schemas/PositiveMoney'
        max_attempts:
          type: integer
          minimum: 1
          maximum: 3
        expires_at: &id002
          type: string
          format: date-time
          pattern: .*(Z|[+-][0-9]{2}:[0-9]{2})$
      required:
      - recipient_ref
      - amount
      - principal_cap
      - max_attempts
      - expires_at
      description: Cap must cover amount; expiry must be future at approval. Workspace allowance remains
        an independent durable control.
...
    Error:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          maxLength: 80
        message:
          type: string
          maxLength: 1000
        request_id: *id001
        details:
          type: array
          items:
            type: object
            additionalProperties: false
            properties:
              field:
                oneOf:
                - type: string
                  maxLength: 200
                - type: 'null'
              reason:
                type: string
                maxLength: 600
            required:
            - field
            - reason
          minItems: 0
          maxItems: 30
        retryable:
          type: boolean
      required:
      - code
      - message
      - request_id
      - details
      - retryable
      description: Codes include INVALID_REQUEST, UNSUPPORTED_SCOPE, UNAUTHORIZED, FORBIDDEN, NOT_FOUND,
        STALE_REVISION, IDEMPOTENCY_CONFLICT, WORKSPACE_BUSY, AUTHORITY_EXPIRED, GUARD_FAILED and DEPENDENCY_UNAVAILABLE.
        No secrets or raw receiver values.
    AcceptedCommand:
      type: object
      additionalProperties: false
      properties:
        command_id: *id001
        state:
          type: string
          enum:
          - accepted
          - held
          - deduplicated
        resource_id:
          oneOf:
          - *id001
          - type: 'null'
        job_id:
          oneOf:
          - *id001
          - type: 'null'
        existing_attempt_id:
          oneOf:
          - *id001
          - type: 'null'
        hold_reasons:
          type: array
          items:
            type: string
            maxLength: 160
          minItems: 0
          maxItems: 12
      required:
      - command_id
      - state
      - resource_id
      - job_id
      - existing_attempt_id
      - hold_reasons
      description: 202 acknowledges durable work only. Held reset/payment work is persisted with reasons;
        it has not dispatched.
...
    BriefCreate:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
          maxLength: 160
          minLength: 1
        text:
          type: string
          maxLength: 4000
          minLength: 1
        family: *id003
        fixture_ref:
          type: string
          enum:
          - checkout_fixture
        terms:
          $ref: '#/components/schemas/PaymentTerms'
      required:
      - title
      - text
      - family
      - fixture_ref
      - terms
...
    Brief:
      type: object
      additionalProperties: false
      properties:
        id: *id001
        task_id: *id001
        revision_id: *id001
        revision:
          type: integer
          minimum: 1
        digest: *id004
        title:
          type: string
          maxLength: 160
        text:
          type: string
          maxLength: 4000
        family: *id003
        fixture_ref:
          type: string
          enum:
          - checkout_fixture
        terms:
          $ref: '#/components/schemas/PaymentTerms'
        created_at: *id002
        updated_at: *id002
      required:
      - id
      - task_id
      - revision_id
      - revision
      - digest
      - title
      - text
      - family
      - fixture_ref
      - terms
      - created_at
      - updated_at
    CompileRequest:
      type: object
      additionalProperties: false
      properties:
        brief_revision_id: *id001
        expected_digest: *id004
      required:
      - brief_revision_id
      - expected_digest
...
    MandateDraftCreate:
      type: object
      additionalProperties: false
      properties:
        task_id: *id001
        brief_revision_id: *id001
        compilation_id: *id001
        terms:
          $ref: '#/components/schemas/PaymentTerms'
      required:
      - task_id
      - brief_revision_id
      - compilation_id
      - terms
...
    ApproveRequest:
      type: object
      additionalProperties: false
      properties:
        expected_draft_digest: *id004
        expected_current_version_id:
          oneOf:
          - *id001
          - type: 'null'
      required:
      - expected_draft_digest
      - expected_current_version_id
...
    MandateVersion:
      type: object
      additionalProperties: false
      properties:
        id: *id001
        mandate_id: *id001
        version:
          type: integer
          minimum: 1
        lifecycle_state:
          type: string
          enum:
          - draft
          - approved
          - expired
          - superseded
          - exhausted
        digest: *id004
        snapshot:
          $ref: '#/components/schemas/MandatePayload'
        created_at: *id002
      required:
      - id
      - mandate_id
      - version
      - lifecycle_state
      - digest
      - snapshot
      - created_at
    Mandate:
      type: object
      additionalProperties: false
      properties:
        id: *id001
        task_id: *id001
        current_version_id:
          oneOf:
          - *id001
          - type: 'null'
        versions:
          type: array
          items:
            $ref: '#/components/schemas/MandateVersion'
          minItems: 0
          maxItems: 50
      required:
      - id
      - task_id
      - current_version_id
      - versions
...
    CheckResult:
      type: object
      additionalProperties: false
      properties:
        id: *id001
        ref: *id005
        check_id:
          type: string
          enum:
          - C01
          - C02
          - C03
        type:
          type: string
          enum:
          - viewport_no_horizontal_overflow
          - cart_total_unchanged
          - keyboard_checkout_reachable
          - api_status
          - api_schema
          - api_total_matches_fixture
          - keyboard_activation
          - accessible_control_name
        outcome:
          type: string
          enum:
          - pass
          - fail
          - error
        observations:
          type: object
          additionalProperties: true
        artifact_refs:
          type: array
          items: *id005
          minItems: 0
          maxItems: 4
        completed_at: *id002
      required:
      - id
      - ref
      - check_id
      - type
      - outcome
      - observations
      - artifact_refs
      - completed_at
...
    EvidenceBundle:
      type: object
      additionalProperties: false
      properties:
        id: *id001
        task_id: *id001
        delivery_id: *id001
        mandate_version_id: *id001
        mandate_digest: *id004
        artifact_digest: *id004
        run_id: *id001
        digest: *id004
        claim_ref: *id005
        manifest_ref: *id005
        results:
          type: array
          items:
            $ref: '#/components/schemas/CheckResult'
          minItems: 3
          maxItems: 3
        artifacts:
          type: array
          items:
            $ref: '#/components/schemas/EvidenceArtifact'
          minItems: 0
          maxItems: 6
        review:
          oneOf:
          - $ref: '#/components/schemas/EvidenceReview'
          - type: 'null'
        review_interaction_id:
          oneOf:
          - *id001
          - type: 'null'
        is_current:
          type: boolean
        created_at: *id002
      required:
      - id
      - task_id
      - delivery_id
      - mandate_version_id
      - mandate_digest
      - artifact_digest
      - run_id
      - digest
      - claim_ref
      - manifest_ref
      - results
      - artifacts
      - review
      - review_interaction_id
      - is_current
      - created_at
...
    BriefList:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Brief'
          minItems: 0
          maxItems: 100
        next_cursor:
          oneOf:
          - type: string
            maxLength: 256
          - type: 'null'
      required:
      - items
      - next_cursor
...